Fixes input issues.
This commit fixes input issues as they pertain to the temperature value.
Prior to this the data type that was expected for the app was an _i16_
which was then converted to a f64 for the calculations. This commit
changes the expected data type to f64 so there are no issues using a
float as the input value or using what could be percieved as a integer.

+++ src/main.rs
@@ -3,7 +3,7 @@ use std::env;
 fn main() {
     let args: Vec<String> = env::args().collect();
     let scale = &args[1];
-    let value: i16 = args[2].trim().parse()
+    let value: f64 = args[2].trim().parse()
         .expect("Please enter number!");
 
     if scale == "f" {
@@ -17,30 +17,30 @@ fn main() {
     };
 }
 
-fn fahrenheit(value: i16) {
+fn fahrenheit(value: f64) {
     println!("Converting {}° fahrenheit.", value);
 
-    let celsius_temp = ((((f64::from(value) - 32.0) * (5.0 / 9.0)) * 100.0).round()) / 100.0;
+    let celsius_temp = ((((value - 32.0) * (5.0 / 9.0)) * 100.0).round()) / 100.0;
     let kelvin_temp = (((celsius_temp + 273.15) * 100.0).round()) / 100.0;
 
     println!("{}° fahrenheit is equal to {}° celsius.", value, celsius_temp);
     println!("{}° fahrenheit is equal to {}° kelvin.", value, kelvin_temp);
 }
 
-fn celsius(value: i16) {
+fn celsius(value: f64) {
     println!("Converting {}° celsius.", value);
 
-    let fahrenheit_temp = ((((f64::from(value) * (9.0 / 5.0)) + 32.0) * 100.0).round()) / 100.0;
-    let kelvin_temp = (((f64::from(value) + 273.15) * 100.0).round()) / 100.0;
+    let fahrenheit_temp = ((((value * (9.0 / 5.0)) + 32.0) * 100.0).round()) / 100.0;
+    let kelvin_temp = (((value + 273.15) * 100.0).round()) / 100.0;
 
     println!("{}° celsius is equal to {}° fahrenheit.", value, fahrenheit_temp);
     println!("{}° celsius is equal to {}° kelvin.", value, kelvin_temp);
 }
 
-fn kelvin(value: i16) {
+fn kelvin(value: f64) {
     println!("Converting {}° kelvin.", value);
 
-    let celsius_temp = (((f64::from(value) - 273.15) * 100.0).round()) / 100.0;
+    let celsius_temp = (((value - 273.15) * 100.0).round()) / 100.0;
     let fahrenheit_temp = ((((celsius_temp * (9.0 / 5.0)) + 32.0) * 100.0).round()) / 100.0;
 
     println!("{}° kelvin is equal to {}° fahrenheit.", value, fahrenheit_temp);
